chore: extract CreateAddRequirementFn
