+++ pkg/jx/cmd/add_app.go
@@ -9,8 +9,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	helmchart "k8s.io/helm/pkg/proto/hapi/chart"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	corev1 "k8s.io/api/core/v1"
@@ -313,223 +311,13 @@ func (o *AddAppOptions) Run() error {
 
 func (o *AddAppOptions) createPR(app string, dir string, version string) error {
 
-	modifyChartFn := func(requirements *helm.Requirements, chart *helmchart.Metadata, values map[string]interface{},
-		templates map[string]string, envDir string) error {
-		// See if the app already exists in requirements
-		found := false
-		for _, d := range requirements.Dependencies {
-			if d.Name == app && d.Alias == o.Alias {
-				// App found
-				log.Infof("App %s already installed.\n", util.ColorWarning(app))
-				if version != d.Version {
-					log.Infof("To upgrade the app use %s or %s\n",
-						util.ColorInfo("jx upgrade app <app>"),
-						util.ColorInfo("jx upgrade apps --all"))
-				}
-				found = true
-				break
-			}
-		}
-		// If app not found, add it
-		if !found {
-			requirements.Dependencies = append(requirements.Dependencies, &helm.Dependency{
-				Alias:      o.Alias,
-				Repository: o.Repo,
-				Name:       app,
-				Version:    version,
-			})
-			appDir := filepath.Join(envDir, app)
-			rootValuesFileName := filepath.Join(appDir, helm.ValuesFileName)
-			err := os.MkdirAll(appDir, 0700)
-			if err != nil {
-				return errors.Wrapf(err, "cannot create app directory %s", appDir)
-			}
-			if o.Verbose {
-				log.Infof("Using %s for app files\n", appDir)
-			}
-			if len(o.ValueFiles) == 1 {
-				// We need to write the values file into the right spot for the app
-				err = util.CopyFile(o.ValueFiles[0], rootValuesFileName)
-				if err != nil {
-					return errors.Wrapf(err, "cannot copy values."+
-						"yaml to app directory %s", appDir)
-				}
-				if o.Verbose {
-					log.Infof("Writing file provided by --%s to %s\n", optionValues, rootValuesFileName)
-				}
-			}
-			// Write the release.yaml
-			var gitRepo, releaseNotesURL, appReadme, description string
-			templatesDir := filepath.Join(dir, "templates")
-			if _, err := os.Stat(templatesDir); os.IsNotExist(err) {
-				if o.Verbose {
-					log.Infof("No templates directory exists in %s", dir)
-				}
-			} else if err != nil {
-				return errors.Wrapf(err, "stat directory %s", appDir)
-			} else {
-				releaseYamlPath := filepath.Join(templatesDir, "release.yaml")
-				if _, err := os.Stat(releaseYamlPath); err == nil {
-					bytes, err := ioutil.ReadFile(releaseYamlPath)
-					if err != nil {
-						return errors.Wrapf(err, "release.yaml from %s", templatesDir)
-					}
-					release := jenkinsv1.Release{}
-					err = yaml.Unmarshal(bytes, &release)
-					if err != nil {
-						return errors.Wrapf(err, "unmarshal %s", releaseYamlPath)
-					}
-					gitRepo = release.Spec.GitHTTPURL
-					releaseNotesURL = release.Spec.ReleaseNotesURL
-					releaseYamlOutPath := filepath.Join(appDir, "release.yaml")
-					err = ioutil.WriteFile(releaseYamlOutPath, bytes, 0755)
-					if err != nil {
-						return errors.Wrapf(err, "write file %s", releaseYamlOutPath)
-					}
-					if o.Verbose {
-						log.Infof("Read release notes URL %s and git repo url %s from release.yaml\nWriting release."+
-							"yaml from app to %s\n", releaseNotesURL, gitRepo, releaseYamlOutPath)
-					}
-				} else if os.IsNotExist(err) {
-					if o.Verbose {
-
-						log.Infof("Not adding release.yaml as not present in chart. Only files in %s are:\n",
-							templatesDir)
-						err := util.ListDirectory(templatesDir, true)
-						if err != nil {
-							return err
-						}
-					}
-				} else {
-					return errors.Wrapf(err, "reading release.yaml from %s", templatesDir)
-				}
-			}
-			chartYamlPath := filepath.Join(dir, helm.ChartFileName)
-			if _, err := os.Stat(chartYamlPath); err == nil {
-				bytes, err := ioutil.ReadFile(chartYamlPath)
-				if err != nil {
-					return errors.Wrapf(err, "read %s from %s", helm.ChartFileName, dir)
-				}
-				chart := helmchart.Metadata{}
-				err = yaml.Unmarshal(bytes, &chart)
-				if err != nil {
-					return errors.Wrapf(err, "unmarshal %s", chartYamlPath)
-				}
-				description = chart.Description
-
-			} else if os.IsNotExist(err) {
-				if o.Verbose {
-					log.Infof("Not adding %s as not present in chart. Only files in %s are:\n", helm.ChartFileName, dir)
-					err := util.ListDirectory(dir, true)
-					if err != nil {
-						return err
-					}
-				}
-			} else {
-				return errors.Wrapf(err, "stat Chart.yaml from %s", dir)
-			}
-			// Need to copy over any referenced files, and their schemas
-			rootValues, err := helm.LoadValuesFile(rootValuesFileName)
-			if err != nil {
-				return err
-			}
-			schemas := make(map[string][]string)
-			possibles := make(map[string]string)
-			if _, err := os.Stat(dir); err == nil {
-				files, err := ioutil.ReadDir(dir)
-				if err != nil {
-					return errors.Wrapf(err, "unable to list files in %s", dir)
-				}
-				possibleReadmes := make([]string, 0)
-				for _, file := range files {
-					fileName := strings.ToUpper(file.Name())
-					if fileName == "README.MD" || fileName == "README" {
-						possibleReadmes = append(possibleReadmes, filepath.Join(dir, file.Name()))
-					}
-				}
-				if len(possibleReadmes) > 1 {
-					if o.Verbose {
-						log.Warnf("Unable to add README to PR for %s as more than one exists and not sure which to"+
-							" use %s\n", app, possibleReadmes)
-					}
-				} else if len(possibleReadmes) == 1 {
-					bytes, err := ioutil.ReadFile(possibleReadmes[0])
-					if err != nil {
-						return errors.Wrapf(err, "unable to read file %s", possibleReadmes[0])
-					}
-					appReadme = string(bytes)
-				}
-
-				for _, f := range files {
-					ignore, err := util.IgnoreFile(f.Name(), helm.DefaultValuesTreeIgnores)
-					if err != nil {
-						return err
-					}
-					if !f.IsDir() && !ignore {
-						key := f.Name()
-						// Handle .schema. files specially
-						if parts := strings.Split(key, ".schema."); len(parts) > 1 {
-							// this is a file named *.schema.*, the part before .schema is the key
-							if _, ok := schemas[parts[0]]; !ok {
-								schemas[parts[0]] = make([]string, 0)
-							}
-							schemas[parts[0]] = append(schemas[parts[0]], filepath.Join(dir, f.Name()))
-						}
-						possibles[key] = filepath.Join(dir, f.Name())
-
-					}
-				}
-			} else if !os.IsNotExist(err) {
-				return errors.Wrap(err, fmt.Sprintf("error reading %s", dir))
-			}
-			if o.Verbose && appReadme == "" {
-				log.Infof("Not adding App Readme as no README, README.md, readme or readme.md found in %s\n", dir)
-			}
-			readme := o.generateReadme(app, version, description, o.Repo, gitRepo, releaseNotesURL, appReadme)
-			readmeOutPath := filepath.Join(appDir, "README.MD")
-			err = ioutil.WriteFile(readmeOutPath, []byte(readme), 0755)
-			if err != nil {
-				return errors.Wrapf(err, "write README.md to %s", appDir)
-			}
-			if o.Verbose {
-				log.Infof("Writing README.md to %s\n", readmeOutPath)
-			}
-			externalFileHandler := func(path string, element map[string]interface{}, key string) error {
-				fileName, _ := filepath.Split(path)
-				err := util.CopyFile(path, filepath.Join(appDir, fileName))
-				if err != nil {
-					return errors.Wrapf(err, "copy %s to %s", path, appDir)
-				}
-				// key for schema is the filename without the extension
-				schemaKey := strings.TrimSuffix(fileName, filepath.Ext(fileName))
-				if schemaPaths, ok := schemas[schemaKey]; ok {
-					for _, schemaPath := range schemaPaths {
-						fileName, _ := filepath.Split(schemaPath)
-						schemaOutPath := filepath.Join(appDir, fileName)
-						err := util.CopyFile(schemaPath, schemaOutPath)
-						if err != nil {
-							return errors.Wrapf(err, "copy %s to %s", schemaPath, appDir)
-						}
-						if o.Verbose {
-							log.Infof("Writing %s to %s\n", fileName, schemaOutPath)
-						}
-					}
-				}
-				return nil
-			}
-			err = helm.HandleExternalFileRefs(rootValues, possibles, "", externalFileHandler)
-			if err != nil {
-				return err
-			}
-		}
-
-		return nil
-	}
 	branchNameText := "add-app-" + app + "-" + version
 	title := fmt.Sprintf("Add %s %s", app, version)
 	message := fmt.Sprintf("Add app %s %s", app, version)
 
-	pullRequestInfo, err := o.createEnvironmentPullRequest(o.DevEnv, modifyChartFn, &branchNameText, &title,
+	pullRequestInfo, err := o.createEnvironmentPullRequest(o.DevEnv, o.CreateAddRequirementFn(app, o.Alias, version,
+		o.Repo, o.ValueFiles, dir),
+		&branchNameText, &title,
 		&message,
 		nil, o.ConfigureGitCallback)
 	if err != nil {
@@ -567,38 +355,4 @@ func (o *AddAppOptions) installApp(name string, chart string, version string) er
 	// Attach the secrets to the name CRD
 
 	return o.OnAppInstall(name, version)
-}
-
-func (o *AddAppOptions) generateReadme(app string, version string, description string, chartRepo string,
-	gitRepo string, releaseNotesURL string, appReadme string) string {
-	var readme strings.Builder
-	readme.WriteString(fmt.Sprintf("# %s\n\n|App Metadata|---|\n", unknownZeroValue(app)))
-	if version != "" {
-		readme.WriteString(fmt.Sprintf("| **Version** | %s |\n", version))
-	}
-	if description != "" {
-		readme.WriteString(fmt.Sprintf("| **Description** | %s |\n", description))
-	}
-	if chartRepo != "" {
-		readme.WriteString(fmt.Sprintf("| **Chart Repository** | %s |\n", chartRepo))
-	}
-	if gitRepo != "" {
-		readme.WriteString(fmt.Sprintf("| **Git Repository** | %s |\n", gitRepo))
-	}
-	if releaseNotesURL != "" {
-		readme.WriteString(fmt.Sprintf("| **Release Notes** | %s |\n", releaseNotesURL))
-	}
-
-	if appReadme != "" {
-		readme.WriteString(fmt.Sprintf("\n## App README.MD\n\n%s\n", appReadme))
-	}
-	return readme.String()
-}
-
-func unknownZeroValue(value string) string {
-	if value == "" {
-		return "unknown"
-	}
-	return value
-
 }
